fix sync with transport id
